+++ shape.py
