+++ simple_numpy.py
