--- v03/.gitignore
@@ -1,2 +1,0 @@
-.idea/
-.vscode/
